+++ app/src/main/java/com/baijiahulian/bjhl_liveplayer/App.java
@@ -2,9 +2,6 @@ package com.baijiahulian.bjhl_liveplayer;
 
 import android.app.Application;
 
-import com.baijiahulian.livecore.LiveSDK;
-import com.baijiahulian.livecore.context.LPConstants;
-
 
 /**
  * Created by yanglei on 16/4/12.
@@ -18,7 +15,5 @@ public class App extends Application {
         CrashHandler crashHandler = CrashHandler.getInstance();
         crashHandler.init(this);
         Thread.setDefaultUncaughtExceptionHandler(crashHandler);
-
-        LiveSDK.init("partnerId", LPConstants.LPDeployType.Product);
     }
 }